+++ apps/playgrounds/vite/vite.config.ts
@@ -1,6 +1,7 @@
 import arkenvVitePlugin from "@arkenv/vite-plugin";
 import reactPlugin from "@vitejs/plugin-react";
-import arkenv, { type } from "arkenv";
+import arkenv from "arkenv";
+import { type } from "arkenv/arktype";
 import { defineConfig, loadEnv } from "vite";
 import tsconfigPaths from "vite-tsconfig-paths";